requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
fastui==0.5.2
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
mistralai==0.1.6
numpy==1.26.4
orjson==3.9.15
pandas==2.2.1
pyarrow==15.0.1
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.9
pytz==2024.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
